ipython notebook ver 2 after 1st review
Added explanation for the choice of decision tree and the splitting points based on the following questions:

Describe the steps you took to implement the final prediction model so that it got an accuracy of at least 80%. What features did you look at? Were certain features more informative than others? Which conditions did you use to split the survival outcomes in the data? How accurate are your predictions?